FadeInの共通化して遷移の方法を変更
