+++ flutter/lib/app/pages/sign_in.dart
@@ -3,6 +3,9 @@ import 'package:flutter/material.dart';
 import 'package:bacchus/repository/provider/auth.dart';
 import 'package:bacchus/domain/entities/user.dart';
 
+import 'package:bacchus/app/pages/home.dart';
+import 'package:bacchus/app/widget/fade_in_route.dart';
+
 class SignInPage extends StatefulWidget {
   SignInPage({Key key, this.setUser}) : super(key: key);
 
@@ -42,7 +45,13 @@ class _SignInState extends State<SignInPage> {
     final user = User(userId, userName);
     await user.addStore();
     widget.setUser(user);
-    Navigator.of(context).pushReplacementNamed('/home');
+    Navigator.pushReplacement(
+      context,
+      FadeInRoute(
+        widget: HomePage(),
+        opaque: true,
+      ),
+    );
   }
 
   @override
